refactor(ort-utils): Split `Extensions.kt` and `Utils.kt`
Similar to 6d34ce5, split `ort-utils` by functionality instead of function
type.
